Remove the need to give IdP/SP handlers names
This was due to poor implementation in the past, where the code should
have been checking the EntityID in the incoming request instead.
EntityIDs are used everywhere now, instead of the IdPHandler.name,
including in URLs.

This means that SP URL's - including the metadata XML URL, and therefore
the SP entity ID - are now different. This is a breaking change, which
will require people to reconfigure any IdP's that handle the SP.