Fix a bug with non-signed requests in the IdPHandler class.
If the service provider isn't signing requests, then the
_make_idp_request_url method turns the query string into a
bytestring before including it in a format string below.

This causes the resulting URL to include the literal characters

b'

preceeding the desired contents of the query string.

This commit removes the call to encode('utf-8') so the url
is generated correctly.